refactor: renamed the packages
I renamed the packages to be more in line with how marcelo has them so merging wont be as much of a hassle
