+++ app/src/main/java/com/example/sightreadingapp/UserProfile.kt
@@ -1,23 +1,12 @@
 package com.example.sightreadingapp
 
-import ProfileSelectionScreen
+import com.example.sightreadingapp.ui.screens.ProfileSelectionScreen
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
 
 // --- Data Model for a User Profile ---
 data class UserProfile(val id: Int, val name: String, val score: Int)
