+++ app/src/main/java/com/example/sightreadingapp/models/Accidentals.kt
@@ -1,4 +1,4 @@
-package com.example.sightreadingapp.data
+package com.example.sightreadingapp.models
 
 enum class Accidentals(val accident: String){
     NONE(""),
