+++ app/src/main/java/com/example/sightreadingapp/models/NoteResourcesAndAnswers.kt
@@ -1,4 +1,4 @@
-package com.example.sightreadingapp.data
+package com.example.sightreadingapp.models
 
 import com.example.sightreadingapp.R
