+++ app/src/main/java/com/example/sightreadingapp/ui/screens/LeaderboardScreen.kt
@@ -1,3 +1,5 @@
+package com.example.sightreadingapp.ui.screens
+
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
