+++ app/src/main/java/com/example/sightreadingapp/ui/screens/ProfileSelectionScreen.kt
@@ -1,3 +1,5 @@
+package com.example.sightreadingapp.ui.screens
+
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
